% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{civic_education}
\alias{civic_education}
\title{civic_education dataset}
\format{
A tibble with 800 rows and 9 variables.
}
\source{
Synthetic data generated by civic.icarm team.
}
\usage{
civic_education
}
\description{
Synthetic civic education outcomes data.
}
\keyword{datasets}
