% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{civic_german_credit}
\alias{civic_german_credit}
\title{civic_german_credit dataset}
\format{
A tibble with 1000 rows and 8 variables.
}
\source{
Synthetic data generated by civic.icarm team.
}
\usage{
civic_german_credit
}
\description{
Synthetic German credit scoring fairness benchmark.
}
\keyword{datasets}
