% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{civic_plot_calibration}
\alias{civic_plot_calibration}
\title{Plot calibration curve}
\usage{
civic_plot_calibration(calibration, title = NULL)
}
\arguments{
\item{calibration}{A civic_calibration from civic_calibrate().}

\item{title}{Optional title.}
}
\value{
A ggplot2 object.
}
\description{
Plot calibration curve
}
