% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{civic_plot_fairness}
\alias{civic_plot_fairness}
\title{Plot group-level fairness metric}
\usage{
civic_plot_fairness(fairness, metric = "acc", title = NULL, ref_line = NULL)
}
\arguments{
\item{fairness}{A civic_fairness from civic_fairness().}

\item{metric}{Character. Column name to plot.}

\item{title}{Optional title.}

\item{ref_line}{Optional numeric reference line.}
}
\value{
A ggplot2 object.
}
\description{
Plot group-level fairness metric
}
\examples{
m <- civic_fit(Species ~ ., iris)
iris2 <- iris
iris2$size <- factor(ifelse(iris2$Sepal.Length>5.8,"large","small"))
f <- civic_fairness(m, iris2, "Species", "size")
civic_plot_fairness(f, metric="acc")
}
