% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{civic_plot_importance}
\alias{civic_plot_importance}
\title{Plot feature importance}
\usage{
civic_plot_importance(explainer, n_features = 15L, title = NULL)
}
\arguments{
\item{explainer}{A civic_explainer from civic_explain().}

\item{n_features}{Max features to display. Default 15.}

\item{title}{Optional plot title.}
}
\value{
A ggplot2 object.
}
\description{
Plot feature importance
}
\examples{
m  <- civic_fit(Species ~ ., iris)
ex <- civic_explain(m)
civic_plot_importance(ex)
}
