% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{civic_plot_roc_groups}
\alias{civic_plot_roc_groups}
\title{Plot per-group ROC curves}
\usage{
civic_plot_roc_groups(eoc_tbl, title = NULL)
}
\arguments{
\item{eoc_tbl}{A tibble from civic_equalized_odds_curve().}

\item{title}{Optional title.}
}
\value{
A ggplot2 object.
}
\description{
Plot per-group ROC curves
}
