% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{civic_voting}
\alias{civic_voting}
\title{civic_voting dataset}
\format{
A tibble with 1000 rows and 10 variables.
}
\source{
Synthetic data generated by civic.icarm team.
}
\usage{
civic_voting
}
\description{
Synthetic civic voting participation data.
}
\keyword{datasets}
