% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{civic_plots}
\alias{civic_plots}
\title{Visualisation functions for civic.icarm}
\value{
A ggplot2 object that can be further customised
  with standard ggplot2 syntax.
}
\description{
A family of ggplot2-based visualisation functions for civic.icarm
model objects. All functions return ggplot2 objects that can be
further customised with standard ggplot2 syntax.
}
