% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civic_fit.R
\name{print.civic_model}
\alias{print.civic_model}
\title{Print a civic_model}
\usage{
\method{print}{civic_model}(x, ...)
}
\arguments{
\item{x}{A civic_model object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print a civic_model
}
