% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civic_fit.R
\name{summary.civic_model}
\alias{summary.civic_model}
\title{Summary of a civic_model}
\usage{
\method{summary}{civic_model}(object, ...)
}
\arguments{
\item{object}{A civic_model object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary of a civic_model
}
